Refactored code Introduced base event parser and the child classes TFEventProfiler and SMTFEventProfiler
